test: mirror the split src module layout
Follow the source split in the public-API tests: move the VCS tests under
metadata/vcs as one file per backend (git, hg, svn, bzr) with the shared
freeze helpers in the module root; split the graph tests into extras,
filter and version beside the graph core; and split the render tests into
one file per format (mermaid, graphviz, json, json_tree, text, rich_text,
freeze) with the shared render helpers in the module root.
